# Optimize OpenSimplex 2D and 3D