--- a/src/open_simplex.rs
+++ b/src/open_simplex.rs
@@ -80,12 +80,12 @@ pub fn open_simplex2<T: Float>(seed: &Seed, point: &::Point2<T>) -> T {
     //Contribution (1,0)
     let dx1 = dx0 - one - squish_constant;
     let dy1 = dy0 - zero - squish_constant;
-    value = value + gradient(seed, xs_floor + one, ys_floor + zero, dx1, dy1);
+    value = value + gradient(seed, xs_floor + one, ys_floor, dx1, dy1);
 
     //Contribution (0,1)
     let dx2 = dx1 + one;
     let dy2 = dy1 - one;
-    value = value + gradient(seed, xs_floor + zero, ys_floor + one, dx2, dy2);
+    value = value + gradient(seed, xs_floor, ys_floor + one, dx2, dy2);
 
     if frac_sum > one {
         xs_floor = xs_floor + one;
@@ -155,19 +155,22 @@ pub fn open_simplex3<T: Float>(seed: &Seed, point: &::Point3<T>) -> T {
     let mut dz0 = point[2] - zb;
 
     let mut value = zero;
+    let c0 = one + two * squish_constant;
+    // let c1 = one - squish_constant;
+
     if in_sum <= one { //We're inside the tetrahedron (3-Simplex) at (0,0,0)
         //Contribution (0,0,0)
         value = value + gradient(seed, xsb, ysb, zsb, dx0, dy0, dz0);
 
         //Contribution (1,0,0)
         let dx1 = dx0 - one - squish_constant;
-        let dy1 = dy0 - zero - squish_constant;
-        let dz1 = dz0 - zero - squish_constant;
+        let dy1 = dy0 - squish_constant;
+        let dz1 = dz0 - squish_constant;
         value = value + gradient(seed, xsb + one, ysb, zsb, dx1, dy1, dz1);
 
         //Contribution (0,1,0)
-        let dx2 = dx0 - zero - squish_constant;
-        let dy2 = dy0 - one - squish_constant;
+        let dx2 = dx1 + one;
+        let dy2 = dy1 - one;
         let dz2 = dz1;
         value = value + gradient(seed, xsb, ysb + one, zsb, dx2, dy2, dz2);
 
@@ -178,19 +181,19 @@ pub fn open_simplex3<T: Float>(seed: &Seed, point: &::Point3<T>) -> T {
         value = value + gradient(seed, xsb, ysb, zsb + one, dx3, dy3, dz3);
     } else if in_sum >= two { //We're inside the tetrahedron (3-Simplex) at (1,11)
         //Contribution (1,1,0)
-        let dx3 = dx0 - one - two * squish_constant;
-        let dy3 = dy0 - one - two * squish_constant;
-        let dz3 = dz0 - zero - two * squish_constant;
+        let dx3 = dx0 - c0;
+        let dy3 = dy0 - c0;
+        let dz3 = dz0 - c0 + one;
         value = value + gradient(seed, xsb + one, ysb + one, zsb, dx3, dy3, dz3);
 
         //Contribution (1,0,1)
         let dx2 = dx3;
-        let dy2 = dy0 - zero - two * squish_constant;
-        let dz2 = dz0 - one - two * squish_constant;
+        let dy2 = dy3 + one;
+        let dz2 = dz3 - one;
         value = value + gradient(seed, xsb + one, ysb, zsb + one, dx2, dy2, dz2);
 
         //Contribution (0,1,1)
-        let dx1 = dx0 - zero - two * squish_constant;
+        let dx1 = dx3 + one;
         let dy1 = dy3;
         let dz1 = dz2;
         value = value + gradient(seed, xsb, ysb + one, zsb + one, dx1, dy1, dz1);
@@ -208,8 +211,8 @@ pub fn open_simplex3<T: Float>(seed: &Seed, point: &::Point3<T>) -> T {
         value = value + gradient(seed, xsb + one, ysb, zsb, dx1, dy1, dz1);
 
         //Contribution (0,1,0)
-        let dx2 = dx0 - zero - squish_constant;
-        let dy2 = dy0 - one - squish_constant;
+        let dx2 = dx1 + one;
+        let dy2 = dy1 - one;
         let dz2 = dz1;
         value = value + gradient(seed, xsb, ysb + one, zsb, dx2, dy2, dz2);
 
@@ -220,19 +223,19 @@ pub fn open_simplex3<T: Float>(seed: &Seed, point: &::Point3<T>) -> T {
         value = value + gradient(seed, xsb, ysb, zsb + one, dx3, dy3, dz3);
 
         //Contribution (1,1,0)
-        let dx4 = dx0 - one - two * squish_constant;
-        let dy4 = dy0 - one - two * squish_constant;
-        let dz4 = dz0 - zero - two * squish_constant;
+        let dx4 = dx0 - c0;
+        let dy4 = dy0 - c0;
+        let dz4 = dz0 - c0 + one;
         value = value + gradient(seed, xsb + one, ysb + one, zsb, dx4, dy4, dz4);
 
         //Contribution (1,0,1)
         let dx5 = dx4;
-        let dy5 = dy0 - zero - two * squish_constant;
-        let dz5 = dz0 - one - two * squish_constant;
+        let dy5 = dy4 + one;
+        let dz5 = dz4 - one;
         value = value + gradient(seed, xsb + one, ysb, zsb + one, dx5, dy5, dz5);
 
         //Contribution (0,1,1)
-        let dx6 = dx0 - zero - two * squish_constant;
+        let dx6 = dx5 + one;
         let dy6 = dy4;
         let dz6 = dz5;
         value = value + gradient(seed, xsb, ysb + one, zsb + one, dx6, dy6, dz6);