Removes most of the duplicate calculations and tries to use previously-computed values as much as possible.

### Benchmarks

**Before**

```
test bench_open_simplex2               ... bench:          26 ns/iter (+/- 1)
test bench_open_simplex2_64x64         ... bench:     104,816 ns/iter (+/- 1,131)
test bench_open_simplex3               ... bench:          52 ns/iter (+/- 3)
test bench_open_simplex3_64x64         ... bench:     169,494 ns/iter (+/- 794)
```

**After**

```
test bench_open_simplex2               ... bench:          25 ns/iter (+/- 1)
test bench_open_simplex2_64x64         ... bench:      99,244 ns/iter (+/- 1,410)
test bench_open_simplex3               ... bench:          50 ns/iter (+/- 2)
test bench_open_simplex3_64x64         ... bench:     163,044 ns/iter (+/- 1,071)
```

Running on Ubuntu 14.04 x64 in VirtualBox.

Rust toolchain:

```
nightly-x86_64-unknown-linux-gnu (default)
rustc 1.11.0-nightly (ad7fe6521 2016-06-23)
```