@bjz 2016-06-27
Nice finds!

I'm guessing some of these things were in there to improve the understanding of the algorithm. Is there a way we could preserve some of that? Perhaps with comments? @Cifram or @amaranth might have some suggestions, seeing as they understand the algorithm more than me.